refactor(rf): Reorganise project structure
