+++ tidy3d/plugins/smatrix/smatrix.py
@@ -1,7 +1,7 @@
 # backwards compatibility support for ``from tidy3d.plugins.smatrix.smatrix import ``
 from __future__ import annotations
 
-from .component_modelers.modal import ComponentModeler
-from .ports.modal import Port
+from tidy3d.components.microwave.component_modelers.modal import ComponentModeler
+from tidy3d.components.microwave.ports.modal import Port
 
 __all__ = ["ComponentModeler", "Port"]
